requirements-dev.txt
jsonrpclib
mock
pytest
pytest-cov
pytest-flakes
ipython
